requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0

pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

supabase==2.0.3
httpx==0.24.1

python-jose[cryptography]==3.3.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
pytest==7.4.3
